+++ packages/react/src/solana/hooks/nft/useMintNFT.ts
@@ -14,7 +14,7 @@ import invariant from "tiny-invariant";
  * import { useProgram, useMintNFT } from "@thirdweb-dev/react/solana";
  *
  * export default function Component() {
- *   const program = useProgram("{{program_address}}");
+ *   const { program } = useProgram("{{program_address}}");
  *   const { mutateAsync: mint, isLoading, error } = useMintNFT(program);
  *
  *   function mintNFT() {