+++ packages/react/src/solana/hooks/nft/useTransferNFT.ts
@@ -18,7 +18,7 @@ export type TransferNFTMutationParams = {
  * import { useProgram, useTransferNFT } from "@thirdweb-dev/react/solana";
  *
  * export default function Component() {
- *   const program = useProgram("{{program_address}}");
+ *   const { program } = useProgram("{{program_address}}");
  *   const { mutateAsync: transfer, isLoading, error } = useTransferNFT(program);
  *
  *   return (