+++ packages/sdk/src/solana/classes/deployer.ts
@@ -13,7 +13,6 @@ import { Registry } from "./registry";
 import {
   findMetadataPda,
   getSignerHistogram,
-  InstructionWithSigners,
   Metaplex,
   sol,
   toBigNumber,
@@ -201,7 +200,7 @@ export class Deployer {
    * const metadata = {
    *   name: "My NFT Drop",
    *   symbol: "NFT",
-   *   itemsAvailable: 5,
+   *   totalSuppply: 5,
    * };
    *
    * const address = await sdk.deployer.createNftDrop(metadata);