+++ packages/sdk/test/solana/nft-drop.test.ts
@@ -1,5 +1,6 @@
 import { NFTDrop } from "../../src/solana";
 import { sdk } from "./before-setup";
+import { Keypair } from "@solana/web3.js";
 import { expect } from "chai";
 
 describe("NFTDrop", async () => {
@@ -54,15 +55,21 @@ describe("NFTDrop", async () => {
     await drop.claimConditions.set({
       maxClaimable: 1,
     });
-    const address = await drop.claim(1);
+    const receiver = Keypair.generate();
+    const address = await drop.claimTo(receiver.publicKey.toBase58(), 1);
 
     unclaimed = await drop.totalUnclaimedSupply();
     claimed = await drop.totalClaimedSupply();
     expect(unclaimed).to.equal(4);
     expect(claimed).to.equal(1);
 
-    const balance = await drop.balance(address[0]);
+    const balance = await drop.balanceOf(
+      receiver.publicKey.toBase58(),
+      address[0],
+    );
     expect(balance).to.equal(1);
+    const nftClaimed = await drop.get(address[0]);
+    expect(nftClaimed.owner).to.equal(receiver.publicKey.toBase58());
   });
 
   it("should update claim condition", async () => {
@@ -84,7 +91,9 @@ describe("NFTDrop", async () => {
   });
 
   it("should get all nfts", async () => {
-    await drop.claim(2);
+    const claimedAddresses = await drop.claim(2);
+    const nftClaimed = await drop.get(claimedAddresses[0]);
+    expect(nftClaimed.owner).to.equal(sdk.wallet.getAddress());
 
     const all = await drop.getAll();
     const claimed = await drop.getAllClaimed();